fix: 소셜 로그인 api  타입 폴더 위치 수정
